+++ samplesheets/studyapps/germline/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GermlineConfig(AppConfig):
+    name = 'samplesheets.studyapps.germline'